Latest Cunqa version implemetation

+++ src/cli/qraise.cpp
@@ -16,6 +16,7 @@
 #include "qraise/simple_conf_qraise.hpp"
 #include "qraise/cc_conf_qraise.hpp"
 #include "qraise/qc_conf_qraise.hpp"
+#include "qraise/qmio_conf_qraise.hpp"
 #include "qraise/infrastructure_conf_qraise.hpp"
 
 #include "logger.hpp"
@@ -44,11 +45,8 @@ void write_sbatch_header(std::ofstream& sbatchFile, const CunqaArgs& args)
 
     if (args.qpus_per_node.has_value()) {
         if (args.n_qpus < args.qpus_per_node) {
-            LOGGER_ERROR("Less qpus than selected qpus_per_node.");
-            LOGGER_ERROR("\tNumber of QPUs: {}\n\t QPUs per node: {}", args.n_qpus, args.qpus_per_node.value());
-            LOGGER_ERROR("Aborted.");
             std::system("rm qraise_sbatch_tmp.sbatch");
-            return;
+            throw std::runtime_error("Less QPUs than qpus_per_node");
         } else {
             sbatchFile << "#SBATCH --ntasks-per-node=" << args.qpus_per_node.value() << "\n";
         }
@@ -56,10 +54,7 @@ void write_sbatch_header(std::ofstream& sbatchFile, const CunqaArgs& args)
 
     if (args.node_list.has_value()) {
         if (args.number_of_nodes.value() != args.node_list.value().size()) {
-            LOGGER_ERROR("Different number of node names than total nodes.");
-            LOGGER_ERROR("\tNumber of nodes: {}\n\t Number of node names: {}", args.number_of_nodes.value(), args.node_list.value().size());
-            LOGGER_ERROR("Aborted.");
-            return;
+            throw std::runtime_error("Different number of node names than total nodes");
         } else {
             sbatchFile << "#SBATCH --nodelist=";
             int comma = 0;
@@ -75,16 +70,15 @@ void write_sbatch_header(std::ofstream& sbatchFile, const CunqaArgs& args)
     }
 
     if (args.mem_per_qpu.has_value() && (args.mem_per_qpu.value()/args.cores_per_qpu > DEFAULT_MEM_PER_CORE)) {
-        LOGGER_ERROR("Too much memory per QPU. Please, decrease the mem-per-qpu or increase the cores-per-qpu.");
-        return;
+        throw std::runtime_error("Too much memory per QPU. Please, decrease the mem-per-qpu or increase the cores-per-qpu.");
     }
 
     if (!args.qc) {
         if (args.mem_per_qpu.has_value() && check_mem_format(args.mem_per_qpu.value())) {
-            sbatchFile << "#SBATCH --mem-per-cpu=" << args.mem_per_qpu.value()/args.cores_per_qpu << "G\n";
+            int mem_per_cpu = (args.mem_per_qpu.value()/args.cores_per_qpu != 0) ? args.mem_per_qpu.value()/args.cores_per_qpu : 1;
+            sbatchFile << "#SBATCH --mem-per-cpu=" << mem_per_cpu << "G\n";
         } else if (args.mem_per_qpu.has_value() && !check_mem_format(args.mem_per_qpu.value())) {
-            LOGGER_ERROR("Memory format is incorrect, must be: xG (where x is the number of Gigabytes).");
-            return;
+            throw std::runtime_error("Memory format is incorrect, must be: xG (where x is the number of Gigabytes).");
         } else if (!args.mem_per_qpu.has_value()) {
             int mem_per_core = DEFAULT_MEM_PER_CORE;
             sbatchFile << "#SBATCH --mem-per-cpu=" << mem_per_core << "G\n";
@@ -102,15 +96,11 @@ void write_sbatch_header(std::ofstream& sbatchFile, const CunqaArgs& args)
     if (check_time_format(args.time))
         sbatchFile << "#SBATCH --time=" << args.time << "\n";
     else {
-        LOGGER_ERROR("Time format is incorrect, must be: xx:xx:xx.");
-        return;
+        throw std::runtime_error("Incorrect time format");
     }
 
-    if (!check_simulator_name(args.simulator)){
-        LOGGER_ERROR("Incorrect simulator name ({}).", args.simulator);
-        return;
-    }
 
+    //sbatchFile << "#SBATCH --profile=all\n";   // Enable comprehensive profiling
     sbatchFile << "#SBATCH --output=qraise_%j\n\n";
     sbatchFile << "unset SLURM_MEM_PER_CPU SLURM_CPU_BIND_LIST SLURM_CPU_BIND\n";
     
@@ -135,8 +125,7 @@ void write_run_command(std::ofstream& sbatchFile, const CunqaArgs& args, const s
             LOGGER_WARN("Personalized noise models only supported for AerSimulator, switching simulator setting from {} to Aer.", args.simulator.c_str());
         }
         if (args.cc || args.qc){
-            LOGGER_ERROR("Personalized noise models not supported for classical/quantum communications schemes.");
-            return;
+            throw std::runtime_error("Personalized noise models not supported for classical/quantum communications schemes.");
         }
 
         if (args.backend.has_value()){
@@ -148,9 +137,7 @@ void write_run_command(std::ofstream& sbatchFile, const CunqaArgs& args, const s
 
     } else if ((!args.noise_properties.has_value() || !args.fakeqmio.has_value()) && (args.no_thermal_relaxation || args.no_gate_error || 
 args.no_readout_error)){
-        LOGGER_ERROR("noise_properties flags where provided but --noise_properties nor --fakeqmio args were not included.");
-        return;
-
+        throw std::runtime_error("noise_properties flags where provided but --noise_properties nor --fakeqmio args were not included.");
     } else {
         if (args.cc) {
             LOGGER_DEBUG("Classical communications");
@@ -164,11 +151,16 @@ args.no_readout_error)){
         }
     }
 
+    if (run_command == "0") { 
+        throw std::runtime_error("Unable to get the proper run command.");
+    }
+
+
     LOGGER_DEBUG("Run command: {}", run_command);
     sbatchFile << run_command;
 }
 
-}
+} // End namespace
 namespace fs = std::filesystem;
 
 int main(int argc, char* argv[]) 
@@ -176,34 +168,50 @@ int main(int argc, char* argv[])
     auto args = argparse::parse<CunqaArgs>(argc, argv, true); //true ensures an error is raised if we feed qraise an unrecognized flag
 
     if (args.infrastructure.has_value()) {
-            LOGGER_DEBUG("Raising infrastructure with path: {}", args.infrastructure.value());
-            fs::path current_dir = fs::current_path();
-            LOGGER_DEBUG("Current dir: {}", current_dir.string());
+        LOGGER_DEBUG("Raising infrastructure with path: {}", args.infrastructure.value());
+        fs::path current_dir = fs::current_path();
+        LOGGER_DEBUG("Current dir: {}", current_dir.string());
 
-            std::ofstream sbatchFile("qraise_sbatch_tmp.sbatch");
-            write_sbatch_file_from_infrastructure(sbatchFile, args);
-            sbatchFile.close();
+        std::ofstream sbatchFile("qraise_sbatch_tmp.sbatch");
+        write_sbatch_file_from_infrastructure(sbatchFile, args);
+        sbatchFile.close();
     } else {
+        if (args.n_qpus == 0 || args.time == "") {
+            LOGGER_INFO("qraise needs two mandatory arguments:\n \t -n: number of vQPUs to be raised\n\t -t: maximum time vQPUs will be raised (hh:mm:ss)\n");
+            std::cout << "\033[32m qraise needs two mandatory arguments: \n\t -n: number of vQPUs to be raised\n\t -t: maximum time vQPUs will be raised (hh:mm:ss)\n \033[0m" << std::endl;
+            return 1;
+        }
         // Setting and checking mode and family name, respectively
         std::string mode = args.co_located ? "co_located" : "hpc";
         std::string family = args.family_name;
         if (exists_family_name(family, constants::QPUS_FILEPATH)) { //Check if there exists other QPUs with same family name
             LOGGER_ERROR("There are QPUs with the same family name as the provided: {}.", family);
             std::system("rm qraise_sbatch_tmp.sbatch");
-            return -1;
+            return 1;
         }
 
         // Writing the sbatch file
         std::ofstream sbatchFile("qraise_sbatch_tmp.sbatch");
-        write_sbatch_header(sbatchFile, args);
-        write_run_command(sbatchFile, args, mode);
+        try {
+            if (args.qmio) {
+                write_qmio_sbatch(sbatchFile, args);
+                sbatchFile.close();
+            } else {
+                write_sbatch_header(sbatchFile, args);
+                write_run_command(sbatchFile, args, mode);
+            }
+        } catch (const std::exception& e) {
+            LOGGER_ERROR("Error writing the sbatch file. Aborting. {}", e.what());
+            std::system("rm qraise_sbatch_tmp.sbatch");
+            return 1;
+        }
         sbatchFile.close();
 
     }
 
     // Executing and deleting the file
     std::system("sbatch qraise_sbatch_tmp.sbatch");
-   // std::system("rm qraise_sbatch_tmp.sbatch");
+    std::system("rm qraise_sbatch_tmp.sbatch");
     
     
     return 0;
